rename host preload -> start (#37422)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37422

Changelog: [Internal]

in this simple change, i renamed the `preload` API, which triggers all the rct instance / javascript runtime creation to `start`. `preload` is a confusing API on the host because the host should be agnostic of when the timing to create its dependencies are, and the consumer should determine when initialization is lazy, early, etc.

Reviewed By: cipolleschi

Differential Revision: D45760583

fbshipit-source-id: bea26d21f950474352280292426a5facbe52b60f